+++ sdk/python/pulumi_kubernetes/apps/v1beta1/Deployment.py
@@ -13,9 +13,10 @@ from ... import tables, version
 
 class Deployment(pulumi.CustomResource):
     """
-    DEPRECATED - This group version of Deployment is deprecated by apps/v1beta2/Deployment. See the
-    release notes for more information. Deployment enables declarative updates for Pods and
-    ReplicaSets.
+    DEPRECATED - apps/v1beta1/Deployment is not supported by Kubernetes 1.16+ clusters. Use
+    apps/v1/Deployment instead.
+    
+    Deployment enables declarative updates for Pods and ReplicaSets.
     
     This resource waits until it is ready before registering success for
     create/update and populating output properties from the current state of the resource.